[Decomp][Item] Cohesion split: extract persistence + enchantment/gems to Item*.cpp
